Avoid running make update-deps twice
Introduce the idea of applied updates and pass through to the
analyze method whether the update is about to be applied.  In this
case, mark the Python frozen dependencies update as already applied,
so applying it later does nothing.  Add an update() method to all
analyzers that apply all updates after discovering them.

Add a new Repository wrapper class to handle branch switching so
that the PullRequester class doesn't know about branches, which
will make some things easier in the future.