Refactor to use Path everywhere
This required adding to_dict() methods to dependencies and updates
because ruamel.yaml doesn't know how to serialize a Path object.
That in turn was a good reason to move the dependency objects into
their own class hierarchy and separate files.

Also factor out finding files of a given name in a tree to avoid
duplicating code between Helm chart and Kustomize resource scanning,
and add a new test for that functionality to maintain complete
coverage.