Drop the neophile branch after making a PR
This serves no purpose since we will refresh the commit from
scratch on the next run, so remove it.  Take advantage of this
change to add better testing for the contents of a pull request
and to move testing of the contents of a commit to a better location
in the flow.