Change the commit message and pull request title
Change to "[neophile] Update dependencies" to make it more obvious
this is an automated commit.  Stop passing a title into the
CommitMessage class and instead use a class variable, since it never
changes.  Adjust the tests to expect the value of the class variable
rather than hard-coding the expected value to make it easy to change
it later.