Support updating pull requests for main
The code to do branch manipulation for updating an existing pull
request has to know what branch is the primary branch for the
repository and was assuming master.  Use main if it exists,
otherwise fall back on master.  (There's a nicer fix that requires
a GitHub API call, but I was lazy and did the simple thing.)