Convert neophile manual to new user guide format
Upgrade to the current documenteer and reformat the neophile manual
to use the new user guide format. Ignore Pydantic cross-references,
which won't ever be satisfied since it doesn't provide intersphinx
information. Stop trying to use the ruamel.yaml intersphinx
information since it's bogus; it only contains references to the
documentation pages, not to the API.

The remaining problems with being unable to resolve Path in
subclasses will be addressed in a follow-on code restructuring.