Use relative imports and no TYPE_CHECKING
Match our other packages by adopting relative imports and dropping
use of TYPE_CHECKING. The latter is too fiddly and doesn't seem worth
the effort.

This introduced a circular dependency between Factory and Processor.
Resolve that by removing the ability for a Factory to create a
Processor and instead create a Processor directly. (This will be
revisited in the future.)

Fix the directory into which the internal API documentation was
being generated.