Refactor factory and underlying objects
Previously, the Factory object methods took a path to a package to
analyze, and the Processor used the Factory to create other objects
as needed. This was different than our other packages and a bit of
an inversion of the factory pattern.

Instead, change the underlying objects to take the root of the
directory to analyze or update as a parameter, make the factory
methods take no parameters, and go back to having the factory
create the processor object.