Update documentation for Helm/Kustomize removal
Remove all the references to Helm and Kustomize support. Update
the configuration instructions to reflect that the Helm chart now
uses camel-case for configuration settings.

Revise the Renovate instructions based on a review of the current
Renovate configuration syntax. Mention that it's often better to use
latest as the tag in Docker compose files used only for testing.