Switch from aiohttp to httpx
Following the rest of our packages, use httpx as the client instead
of aiohttp, and use respx to mock httpx responses instead of
aioresponses.

neophile is, strictly speaking, a sync application, and there's no
compelling reason to use an async client internally. But the gidgethub
library is async-only and is our standard across other applications,
so remain async so that we can use that library.