Rework command-line interface
Move the --update and --pr work of neophile analyze into a new
neophile update command, turning neophile analyze into a simpler
command that only outputs the result of the analysis. Introduce a
new neophile check command that's the same as neophile analyze but
prints messages to standard error and exits with a non-zero status
if dependencies aren't up to date.