Redo how GitHub tokens are used
Stop embedding the GitHub token in the remote URL for Git repositories
and instead assume Git actions will be authenticated externally somehow.
Get the GitHub token for GitHub REST API operations from GITHUB_TOKEN
instead of NEOPHILE_GITHUB_TOKEN so that it will work from inside
GitHub Actions runs with the temporary token for the action.

Add a GitHub Actions job to update pre-commit dependencies using this
new facility, and back down a pre-commit dependency so that it will
have something to do.