Redo GitHub actor handling for PRs
If NEOPHILE_GITHUB_EMAIL is set, do not try to use the /user
endpoint to get the name and email address, and instead use that
email address and NEOPHILE_GITHUB_USER (neophile by default) for
the name portion. This allows PRs to be created by GitHub Actions,
where the /user endpoint is not available.