Authenticate as a GitHub app
Rather than using a personal access token, authenticate as a
GitHub app, installed into the GitHub repository for which neophile
is trying to create a PR. For GitHub tag inventory requests, use an
anonymous client, since that should be public information.

This eliminates the last use of github_user as a username, so replace
that configuration with commit_name and commit_email and clarify that
they're only used for Git commits.

Update the periodic GitHub Action to update pre-commit dependencies
to use the new neophile organization-wide GitHub app via organization
secrets.