Update documentation
Rework the user guide documentation for the new way of using
neophile, deleting most of the previous documentation and adding
recommendations for the new GitHub Actions setup.

Fix all redirects across the documentation.

Enable dependabot updates for Python dependencies now that this is
a PyPI package instead of one with frozen dependencies.